Divides main.js by modules
